+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/PlayerMixin.java
@@ -1,0 +1,39 @@
+package insane96mcp.iguanatweaksreborn.mixin;
+
+import insane96mcp.iguanatweaksreborn.module.Modules;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(Player.class)
+public abstract class PlayerMixin extends LivingEntity {
+	@Shadow
+	public int experienceLevel;
+
+	protected PlayerMixin(EntityType<? extends LivingEntity> type, Level level) {
+		super(type, level);
+	}
+
+	@Shadow
+	public abstract boolean isSpectator();
+
+	@Inject(at = @At("RETURN"), method = "getXpNeededForNextLevel", cancellable = true)
+	private void xpBarCap(CallbackInfoReturnable<Integer> callback) {
+		int exp = Modules.experience.playerExperience.getBetterScalingLevel(this.experienceLevel);
+		if (exp != -1)
+			callback.setReturnValue(exp);
+	}
+
+	@Inject(at = @At("HEAD"), method = "getExperienceReward", cancellable = true)
+	private void getExperiencePoints(Player player, CallbackInfoReturnable<Integer> callback) {
+		int exp = Modules.experience.playerExperience.getExperienceOnDeath((Player) (Object) this);
+		if (exp != -1)
+			callback.setReturnValue(exp);
+	}
+}